Add OpenCV back to CMakeLists
OpenCV is needed for the AR detection code, and will likely be needed
for other parts of the project as well, so I've added it back to the
CMake build. OpenCV isn't too hard to build on GNU/Linux systems, and
I think there's a pre-compiled version you can install on Windows
systems.