Add ar subdirectory for AR tag detection code
I added ar as a subdirectory because I need a separate binary for
testing AR detection to be compiled. This binary doesn't affect the
rest of the code, all of its code will be contained in
src/ar/ARTester.cpp, and I'll make sure it compiles before pushing
anything.

I also feel like having the AR stuff in a subdirectory kind of helps
to separate it from the rest of the project.