Added the SyntheticLidar class
Synthesizes "perfect" Lidar data in the shape of either a circle of radius
r or a square of width (and height) distance. These can be respectively
achieved by calling SyntheticLidar::nextPolarCircle(double radius) and
SyntheticLidar::nextPolarSquare(double dimension) on an instance.

It should be noted that frames only sample data when either of the
two methods listed above is called explicitly by the client program,
so their size is highly dependent on the frequency at which they are.
The SyntheticLidar::getLastFrame() method can then provide the vector
of all Polar2D coordinates that were created by the aforementioned
calls, in the most recently completed frame (full revolution).
If it takes more than one complete revolutions worth of time to
call either of the next Polar2D methods twice framing will not
be accurate.