Create somewhat working version of detection.
Still needs a lot of improvements, but this commit contains a working
version of the detection and orientation calculation process.

Areas that now need fine-tuning: filtering out what are and are not
tags in the detection process. Probably will try to read the tag data
and if it cannot be done, shape is probably not a tag