Create first working version of AR Tag detection (#12)
* Add OpenCV back to CMakeLists

OpenCV is needed for the AR detection code, and will likely be needed
for other parts of the project as well, so I've added it back to the
CMake build. OpenCV isn't too hard to build on GNU/Linux systems, and
I think there's a pre-compiled version you can install on Windows
systems.

* Add temporary fix to get project to compile

* Add ar subdirectory for AR tag detection code

I added ar as a subdirectory because I need a separate binary for
testing AR detection to be compiled. This binary doesn't affect the
rest of the code, all of its code will be contained in
src/ar/ARTester.cpp, and I'll make sure it compiles before pushing
anything.

I also feel like having the AR stuff in a subdirectory kind of helps
to separate it from the rest of the project.

* Add somewhat working version of the tester/detector

* Add subdirectories

* Fix ObstacleSim so that the project compiles

* Add AR tag detection headers

* Add findAngle function

* Add intrinsic camera parameters

* Create slightly working tester program

* Add field for pitch

* add getter method in tag header file

* Add Tag constructor

* added getCorners and calcOrientation

* fix blank

* Fix bugs with and finish calcOrientation

* getPitch Yaw and Roll

* Fix issue with solvePnP

* Fix issue with point ordering

* add code to draw lines along tag

* add code to draw lines along tag

* Improve detection process

Improved detection by only returning outside contours, preventing
detection of smaller shapes inside the AR tag.

* add getCenter and supporting methods

* Fix bugs with Tag

* Create somewhat working version of detection.

Still needs a lot of improvements, but this commit contains a working
version of the detection and orientation calculation process.

Areas that now need fine-tuning: filtering out what are and are not
tags in the detection process. Probably will try to read the tag data
and if it cannot be done, shape is probably not a tag

* Add documentation.

* Fix some issues with tabs and spaces

* Add detailed installation instructions to README

* Fix most issues mentioned in pull request review

Only issue that is not fixed so far is the error code issue, as I'm
not sure yet how to implement that. Will add a fix for that later.

* Remove the use of exceptions as per pull request review

* Fix more style-related issues mentioned in code review