Added real lidar code to read from the Hokuyo URG sensor
Uses the URG SDK (URGC 1.2.5) to read frames from the sensor.
While many of the methods are written similarly to SyntheticLidar the class is NOT a drop in replacement; bigger code changes need to occur to integrate.
The lidar should always be opened (URGLidar::open) before it is read from and closed (URGLidar::close) before the program shuts down.