Replace PointXY with pose_t
We want our navigation goals to include orientation, so we need to
specify them as a pose_t instead of PointXY.

This commit also removes a lot of code from the previous attempt at
autonomous navigation, which I don't really understand and which didn't
quite seem to work.

This commit ALSO makes "calibrated" default to true, so we skip the
calibration phase at the beginning of navigation. (This saves time, and
it doesn't seem to harm navigation performance.)