Implement A* planning for obstacle avoidance
This also implements showing a visualization window so we can visualize
what the rover is planning to do.

The current code replans once every two seconds and in the mean time
does a hacky kind of pure pursuit.

Due to the infrequent replanning, sometimes the rover crashes before
replanning is triggered. In a future PR, we should fix this by verifying
the plan at each timestep using the most recent lidar scans.