EKF Pose Estimator (#58)
Adds an EKF implementation. This had some overlap with the existing KF, so I created a KalmanFilterBase class, which implements the common functionality, which KF and EKF both extend.

Added a pose estimator using the new EKF. The old pose estimator using a KF was renamed to LinearPoseEstimator. I've kept it instead of deleting it because it's never been tested on real robots.

I moved the static utility methods from KalmanFilter to StateSpaceUtil.h, since they're not specific to kalman filters.

Added a wheel base width constant to Constants.h

I also changed the calibration procedure. Now, the estimator is initialized with a single gps reading, which reduces the start-up time. The filter estimates and accounts for the noise in this measurement, so it can tolerate the inaccuracy from a single data point.

It's difficult to gauge how well it'll do on a real robot but it works pretty well in simulations. It's significantly more accurate than either gps or odometry alone.