Improve collision avoidance for planning
(1) We only switch to a new plan if its cost is significantly lower than
the current plan's cost.

(2) We stop the robot if the plan collides with an object near the
robot.

(3) As a refactor, the "should_replan" boolean is removed in favor of
just setting the plan cost to INFINITE_COST.