Visualize landmarks in PlanViz
This commit also demonstrates how we can use the PoseArray message type
to send a long list of points all at once. It also starts to think about
how we'll eventually use transform frames.

When we move to 3D, we're going to have to come up with a better way to
associate landmark location information with the ID of the landmark.
Right now we use that the z-value is zero, which might not work well
once we have landmarks where the z-value is important.