Restructure data written by calibration program
The OpenCV file storage framework can only read/write data structures
as nested dictionaries under a key; writing to the root node of the
file is difficult (while reading is easy). This commit changes the
calibration program to write intrinsic parameters nested under an
"intrinsic_params" key, so it can be read/written using the
CameraParams serialization functions.

Additionally, other data required for camera configuration as well as
explanatory comments are now written to the file.