Attempt to deal with having no heading information from the GPS
We can check whether we have new GPS information via gps.norm(). This
also increases the std_dev of the heading for the EKF to a large value
(essentially saying that this provides no information).

This is not enough to get navigation to work without a heading sensor.
We'll need to implement some smarter algorithms.