Implement incremental PID interface for arm motors
The base station can now send JSON packets of the format
{'type':'motor','motor':'arm_base','incremental PID speed':1.0}

This is similar to PWM control except that it uses feedback from the
encoders to keep the arm moving at a consistent speed (regardless of
external factors such as carrying a weight, for example).

This code handles tracking whether each mode is using PID or PWM control
and switching the motor board modes as necessary.

It also implements a little bit of the inverse kinematics version of
this: we also accept JSON packets of the format
{'type':'motor','motor':'IK_X','incremental IK speed':1.0}
except that we'll return an error saying we haven't implemented this
yet.