Monday changes from rover testing (#111)
* Send status data back to mission control

* Implement incremental PID interface for arm motors

The base station can now send JSON packets of the format
{'type':'motor','motor':'arm_base','incremental PID speed':1.0}

This is similar to PWM control except that it uses feedback from the
encoders to keep the arm moving at a consistent speed (regardless of
external factors such as carrying a weight, for example).

This code handles tracking whether each mode is using PID or PWM control
and switching the motor board modes as necessary.

It also implements a little bit of the inverse kinematics version of
this: we also accept JSON packets of the format
{'type':'motor','motor':'IK_X','incremental IK speed':1.0}
except that we'll return an error saying we haven't implemented this
yet.

* Fix duplicate variable name

* More useful log message

* Use carriage return to save console space

* Increase elbow and forearm PWM scales

It seems the arm team made some changes and now these motors need a bit
more power to move the joints. We're actually currently maxing out the
PWM for the elbow when moving upward, so we may need some firmware
changes? I'm not sure if this will prevent us from lifting up heavy
objects.

* Remove 90 usec from sleep to better hit 10Hz

* Fix incremental PID sign for elbow

* Add .motion config files and update example cameras.rules

* Use log() for planning messages

* Increase lidar min range

* Run planning even when Globals::AUTONOMOUS is false

This lets us visualize whether everything is working together properly.
Now that planning has a separate thread, we don't have to worry so much
about computation.

I also pulled out the map-building code using a boolean USE_MAP. As we
expected, the current version uses a bit too much computation so it
defaults to false.

* Add comment explaining 90us change

Co-authored-by: Husky <husky@jetson>