Merge no-gps stuff and other fixes from on-rover testing (#113)
* Attempt to deal with having no heading information from the GPS

We can check whether we have new GPS information via gps.norm(). This
also increases the std_dev of the heading for the EKF to a large value
(essentially saying that this provides no information).

This is not enough to get navigation to work without a heading sensor.
We'll need to implement some smarter algorithms.

* Update simulator

* Added some commands for levels 1-3

For better modularity I've made it into "commands", so that way once we get  good navigation info we don't have to change too much code. I think it's easier for understanding the logic too

* Minor fix

* Implemented DriveThroughGate command

* Added some code to test traversals

* Fix newline on log message

* Implement cmd_vel integration for real-world readOdom

* Oops I forgot to test compiling real_world_interface

* Tweak wheel parameters for better odometry

* Cleanup of constants

* Fix bug with incremental PID

* Implement command to toggle autonomous mode; increase manual drive speeds

* Tune PID coeffs for arm_base and elbow

* Increase max PWM for the wheels

* Fix broken test

Co-authored-by: Abhay D <a.deshpande012@gmail.com>
Co-authored-by: Husky <husky@jetson>