Navigate to gates and traverse them without compass (#116)
* Attempt to deal with having no heading information from the GPS

We can check whether we have new GPS information via gps.norm(). This
also increases the std_dev of the heading for the EKF to a large value
(essentially saying that this provides no information).

This is not enough to get navigation to work without a heading sensor.
We'll need to implement some smarter algorithms.

* Update simulator

* Added some commands for levels 1-3

For better modularity I've made it into "commands", so that way once we get  good navigation info we don't have to change too much code. I think it's easier for understanding the logic too

* Minor fix

* Implemented DriveThroughGate command

* Added some code to test traversals

* Fix newline on log message

* Implement cmd_vel integration for real-world readOdom

* Oops I forgot to test compiling real_world_interface

* Tweak wheel parameters for better odometry

* Added command to get to the gates, tweaked ekf parameters

* Minor fix

* Respect the autonomous flag

* Fix build error

* Stop driving when autonomy mode changes

* Fix bug with preprocessor conditionals

* Fix bug with array access at -1

For posts (URC legs with only one AR marker) we were trying to access
landmarks[-1] which led to weird behavior.

* Update simulator to latest commit

* Add ability to reset Commands

* Rewrite most of DriveToGateNoCompass

I was getting weird bugs with covariance matrix calculation, so I threw
up my hands and we're just using difference between the first and last
GPS measurement.

* Update submodule

Co-authored-by: Nathan Hatch <nathanhatch297@gmail.com>
Co-authored-by: Husky <husky@jetson>