Clean up various things (#118)
* Add external tag files to Doxygen documentation

* Fix compiler warnings in gpsd_test.cpp

* Fix indentation in gpsd_test.cpp

* Remove unused EKFSlam files

* Fix compiler warnings about narrowing conversion

* Move rover configuration files to separate directory

Additionally, I added a README.md for this directory with a table that
contains a description of these files and where they should go on the
rover. Hopefully this can help to keep track of them and keep them
more organized.

* Move some values into variables in CAN/GPS setup script

* fixup! Add external tag files to Doxygen documentation

* Use ${} syntax for variables in enable_can_and_GPS.sh

* Fix some CI stuff

Moved tests-docker-action to the .github directory rather than the
top-level directory, just for organization. Additionally, made sure
submodules are checked out when the repository is checked out at the
beginning of each CI action.

* fixup! Fix some CI stuff

* Remove unused files

* Remove unused mapping subdirectory