Fix initializer list ordering warnings (-Wreorder)
Fields are initialized in the order they are specified in the class
declaration, so this compiler warning is supposed to warn about the
order of fields in the initializer list being possibly misleading;
this can cause problems if the values of previous fields are used in
the initialization of later ones. I don't think we actually do this
anywhere in our code but it seems like a good practice to address the
warning just so there are no surprises later.