Enable and fix compiler warnings (#119)
* Enable all compiler warnings

* fix #pragma in LidarVis.h

* Remove unused variables

* Explicitly initialize some variables

* Catch exceptions by reference instead of by value

* Fix comparisons between signed and unsigned values

In most cases, this was caused by using a signed integer type for a
loop/array index; we probably ought to use size_t in these cases just
so there's no unexpected behavior from comparing signed/unsigned
types. In some other cases, I explicitly cast the signed value to
unsigned before comparing; we might need to consider whether the
signed value actually needs to be signed in these cases.

* Return something from ParseCommandLineOptions

* Fix initializer list ordering warnings (-Wreorder)

Fields are initialized in the order they are specified in the class
declaration, so this compiler warning is supposed to warn about the
order of fields in the initializer list being possibly misleading;
this can cause problems if the values of previous fields are used in
the initialization of later ones. I don't think we actually do this
anywhere in our code but it seems like a good practice to address the
warning just so there are no surprises later.

* Add some fixes to IK code

* Clean up some unsigned stuff

* Mark Eigen includes as system includes so warnings aren't reported

Co-authored-by: Nathan Hatch <nathanhatch297@gmail.com>