Add option to undistort images in the AR tag Detector
This may help with detection; the camera on the rover has a lot of
barrel distortion from its lens and so it could potentially be unable
to approximate the tags with quadrilaterals if the sides appear too
curved. The option is true by default and will undistort the image
when it is passed into the detectTags() method; the original will not
be modified, and we use cv::initUndistortRectifyMap/cv::remap to make
sure it is performant.