Implement using pose graph optimization to estimate landmark locations (#126)
Previously we used a filter, which essentially took a moving average of
the location estimates over time. However, now that the pose graph
solves may suddenly shift the robot's estimated location, just taking
the average of the location estimates is not a good idea.

Fortunately, the pose graph itself can solve for landmark locations if
we give it the data. This commit implements that.

Co-authored-by: Evan Straw <evan.straw99@gmail.com>