Move ROS publishing to a separate .cpp file
I thought that including rclcpp.h was slowing down compilation for
Rover.cpp and Autonomous.cpp, so I decided to move that into a separate
file rospub.cpp so that we don't need to recompile the publishing logic
every time we e.g. add a print statement to Autonomous.cpp for
debugging.

In the end, I'm not sure this actually improved compile times by much,
but I think it's a nice refactor regardless. Anything that shortens the
Autonomous.cpp file is a good thing, IMO....