Improve GPS error messages and status checks
Something different about ubuntu 20.04 seems to make it so that
'sudo service gpsd stop' doesn't actually stop gpsd. We must also run
'systemctl stop gpsd.socket' to stop gpsd.

Of course, it would be nice if we could just the systemctl version of gpsd
and not manually run it ourselves. But for some reason the systemctl version
doesn't actually publish data when the GPS is plugged in. Someone should
investigate this at some point, but for now, this works.