Set a default value for CMAKE_BUILD_TYPE in config
This code will check if the CMAKE_BUILD_TYPE variable is set and if
not, set it to "RelWithDebInfo". This lets us still specify which type
we would like to build while falling back on a reasonable default. The
config will also print a message at the beginning stating which build
type is being built, and the flag to change it.