Fix GPS bug and increase flexibility of SLAM solver vs. sensor availability (#133)
* WIP: Try to get CV running on rover again

* Only print landmarks if they are visible

* Improve GPS error messages and status checks

Something different about ubuntu 20.04 seems to make it so that
'sudo service gpsd stop' doesn't actually stop gpsd. We must also run
'systemctl stop gpsd.socket' to stop gpsd.

Of course, it would be nice if we could just the systemctl version of gpsd
and not manually run it ourselves. But for some reason the systemctl version
doesn't actually publish data when the GPS is plugged in. Someone should
investigate this at some point, but for now, this works.

* Gitignore vim swap files

* Revert check for no GPS data

Waiting for GPS data is a way of ensuring that the pose graph does not
grow too quickly. Changing our localization stack to revert to odometry
when there is no GPS data is a larger change that should be in a different
PR.

* Add USE_SLAM option

Setting this to false is useful if the AR tags might move around while the
rover is watching. Otherwise the SLAM optimization does weird things.

* Start at leg idx 0

* Fix some post visibility bugs

* Reduce log verbosity

* WIP: Different approach

The goal is to support two use cases:
1. if GPS is not available, still run SLAM with the camera images
2. add a boolean for whether AR tags might move

* Add log message for pose graph solution

* Bugfix: plan_cost was zero for invalid plan

The initial empty plan was getting given a plan_cost of zero, so no future
plan was able to do better and we never got a plan. This commit fixes that,
adds some debug messages, and factors out a planning-related chunk of
autonomyIter() into a separate method getDriveTargetFromPlan().

* Bugfix: plan_size = 0 is not necessarily invalid

So we should still check for drift and replan if necessary.

* Build with debug symbols by default

* Bugfix: Should sometimes do pose graph solve even when we haven't moved

* Start at origin; set MOVING_LANDMARKS = false by default

* Revert to simulator_legs.txt

* Fix comment on zero_legs.txt

* Remove -g flag

Instead people can just say cmake -DCMAKE_BUILD_TYPE=Debug ../src

* Add comments about GPS startup script

* Update submodule

Co-authored-by: Husky <husky@jetson>