Change rospub.cpp to use a WebSocket server instead of ROS
ROS is currently only used for networking between the rover/2D
simulator and PlanViz. It currently causes a few problems by being a
huge collection of libraries (of which we only use a couple) and being
generally unhygienic (headers/libraries aren't placed at standard
system directories but at a non-standard path, and you're supposed to
use a script that sets LD_LIBRARY_PATH and other environment variables
to be able to use ROS; this is a big headache for setup for new users
and on our CI machines). It will also pose some difficulties in the
future when we make the Rover program cross-platform; there is no good
package for ROS on a non-Debian system, and the docs suggest building
from source instead, which will be a huge headache.

In contrast, websocketpp and Boost are packaged and readily available
on all platforms, much less bloated than ROS for the use case, and we
are already intending to use WebSocket anyway for our network protocol
in the future. This commit is a first step in replacing the
ROS-dependent parts of our code with WebSockets, and a future commit
will remove ROS as a dependency entirely.

NOTE: PlanViz has NOT yet been updated to use WebSocket so until a
future commit does this, PlanViz WILL NOT WORK.