Fix segfault in GlobalMapTest on my machine
Seems like p2 is empty for whatever reason on my machine, and so since
the code assumes that p2 is at least the size of p1, it tries to read
at a given index out of p2 but segfaults because it's empty. This
change prevents it from segfaulting by making the assumption that
p2.size() >= p1.size() explicit; the test fails, but doesn't segfault,
so the test code can move on to other tests.

This isn't a huge issue since it doesn't happen on the CI, but it's
kind of weird that this happens; maybe I should check it out later.