Get rid of `WITH_CV` option in CMakeLists.txt
It's not really going to be practical for us to stub out CV-related
code anymore, since we require OpenCV in all cases anyway due to the
world interface (and trying to find a solution to allow that to be
stubbed out would be harder than just requiring OpenCV on all
systems). Ubuntu has a sufficient version of OpenCV packaged in its
catalog anyway so we won't have the problem that inspired the stubbing
in the first place (we were trying to run the rover for a demo and
couldn't get it to compile because OpenCV was out of date).