Add a condition variable to allow immediately stopping thread
Rather than having the repeat thread wait for a set amount of time,
now we wait on a condition variable for a set amount of time, waking
up if either the time expires or the thread is notified via the
condition variable. This will allow us to signal that the thread
should unblock when we need to kill it (at disconnection or
destruction).

(I hope I am using all of this correctly.)