Autonomous Task (#322)
* Create AutonomousTask to handle autonomous navigation

Use DriveToWaypointCommand functionality to execute an autonomous task.

Add necessary constants to Constants.h

Start autonomous task from MissionControlProtocol.cpp

* fix autonomous: SQUASH ME

* pretty much working + squash me

* Add message validation back in

* Run clang-format

* Use correct json message names

* MissionControl: Accept both floats and unsigned numbers in autonomous

JavaScript truncates any value such as 0.0 or 1.0 to 0 and 1,
respectively. This fails validation because we looks for floats and
those come across as unsigned ints. To fix this, we check if the
incoming number is either a float or an unsigned number. Both are
accepted.

* Change rot order

* format

* wip: Debugging and whatnot

* Improve logging, add slow distance threshold to dtwcommand

* Move autonomous constants to Constants.cpp to match new impl

* format

* Remove extraneous logging

---------

Co-authored-by: Abhay D <a.deshpande012@gmail.com>